Cleanup FilledButton
